renaming to rest_handlers
