+++ litmus-portal/graphql-server/pkg/rest_handlers/file_handler.go
@@ -1,4 +1,4 @@
-package handlers
+package rest_handlers
 
 import (
 	"log"
