+++ litmus-portal/graphql-server/pkg/self-deployer/start.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 
 	clusterHandler "github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/cluster/handler"
-	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/handlers"
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/k8s"
+	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/rest_handlers"
 
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/graph/model"
 )
@@ -35,7 +35,7 @@ func StartDeployer(projectID string) {
 		log.Print("SELF CLUSTER REG FAILED[DB-REG] : ", err)
 	}
 
-	response, statusCode, err := handlers.GetManifest(resp.Token)
+	response, statusCode, err := rest_handlers.GetManifest(resp.Token)
 	if err != nil {
 		log.Print("ERROR", err)
 	}
